# openapi3.yaml
openapi: 3.0.1
info:
  title: config-server
  description: This is a config server that provides the means to manage all the configurations
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
paths:
...
  /schema:
    get:
      parameters:
        - name: id
          in: query
          description: The id of the requested schema
          required: true
          schema: 
            $ref: '#/components/schemas/schemaId'
        - name: shouldDereference
          in: query
          description: should the server bundle all refs into one schema
          schema:
            type: boolean
      operationId: getSchema
      summary: returns the requested schema
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
        '400':
          $ref: '#/components/responses/400BadRequest'
        '404':
          $ref: '#/components/responses/404NotFound'
        '500':
          $ref: '#/components/responses/500InternalServerError'
  /schema/tree:
    get:
      operationId: getSchemasTree
      summary: return a tree representation of all the schemas
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/schemaTree"
        '400':
          $ref: '#/components/responses/400BadRequest'
        '500':
          $ref: '#/components/responses/500InternalServerError'
...
security: []
components:
  responses:
    400BadRequest:
      description: BadRequest
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
    404NotFound:
      description: Not Found - If client does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
...
    500InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
...
  schemas:
    error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
...
    schemaId:
      type: string
      pattern: "^https:\\/\\/mapcolonies\\.com\\/.+$"
      example: "https://mapcolonies.com/common/db/v1"
...
    schemaTree:
      type: array
      example: 
        - name: common
          children:
           - name: boilerplate
             children:
               - name: v1
                 url: https://mapcolonies.com/common/boilerplate/v1
               - name: v2
                 url: https://mapcolonies.com/common/boilerplate/v2
               - name: v3
                 url: https://mapcolonies.com/common/boilerplate/v3
           - name: db
             children:
               - name: v1
                 url: https://mapcolonies.com/common/db/v1
               
      items:
        oneOf:
          - $ref: "#/components/schemas/schemaTreeItem"
          - $ref: "#/components/schemas/schemaTreeDir"
    schemaTreeItem:
      type: object
      properties: 
        name: 
          type: string
        id:
          type: string
    schemaTreeDir:
      type: object
      properties:
        children:
          $ref: "#/components/schemas/schemaTree"
        name: 
          type: string
...